debug code added + some imput added
